// backend/server/api/models/http.code.ts
export enum HttpCode {
    OK = 200,
    Created = 201,
    Accepted = 202,
    NoContent = 204,
    BadRequest = 400,
    Unauthorized = 401,
    Forbidden = 403,
    NotFound = 404,
    Conflict = 409,
    Gone = 410,
    PreConditionFailed = 412,
    ToManyRequest = 429,
    InternalServerError = 500,
    ServiceUnavailable = 503
}